+++ source/mmQuantize.cpp
@@ -69,9 +69,9 @@ bool Quantize::initialize(Context* ctx, std::string app, int argc, char* argv[])
 				cxxopts::value<std::string>())
 			("maxPos", "max corner of vertex position bbox, a string of three floats. Computed of not set.",
 				cxxopts::value<std::string>())
-			("minUv",  "min corner of vertex texture coordinates bbox, a string of three floats. Computed of not set.",
+			("minUv", "min corner of vertex texture coordinates bbox, a string of three floats. Computed of not set.",
 				cxxopts::value<std::string>())
-			("maxUv",  "max corner of vertex texture coordinates bbox, a string of three floats. Computed of not set.",
+			("maxUv", "max corner of vertex texture coordinates bbox, a string of three floats. Computed of not set.",
 				cxxopts::value<std::string>())
 			("minNrm", "min corner of vertex normal bbox, a string of three floats. Computed of not set.",
 				cxxopts::value<std::string>())
@@ -83,6 +83,8 @@ bool Quantize::initialize(Context* ctx, std::string app, int argc, char* argv[])
 				cxxopts::value<std::string>())
 			("outputVar", "path to the output variables file.",
 				cxxopts::value<std::string>())
+			("useFixedPoint", "internally convert the minPos and maxPos to fixed point 16.",
+				cxxopts::value<bool>())
 			;
 
 		auto result = options.parse(argc, argv);
@@ -95,7 +97,7 @@ bool Quantize::initialize(Context* ctx, std::string app, int argc, char* argv[])
 		}
 		//	
 		if (result.count("inputModel"))
-			inputModelFilename = result["inputModel"].as<std::string>();
+			_inputModelFilename = result["inputModel"].as<std::string>();
 		else {
 			std::cerr << "Error: missing inputModel parameter" << std::endl;
 			std::cout << options.help() << std::endl;
@@ -103,7 +105,7 @@ bool Quantize::initialize(Context* ctx, std::string app, int argc, char* argv[])
 		}
 		//
 		if (result.count("outputModel"))
-			outputModelFilename = result["outputModel"].as<std::string>();
+			_outputModelFilename = result["outputModel"].as<std::string>();
 		else {
 			std::cerr << "Error: missing outputModel parameter" << std::endl;
 			std::cout << options.help() << std::endl;
@@ -111,86 +113,89 @@ bool Quantize::initialize(Context* ctx, std::string app, int argc, char* argv[])
 		}
 
 		if (result.count("outputVar"))
-			outputVarFilename = result["outputVar"].as<std::string>();
+			_outputVarFilename = result["outputVar"].as<std::string>();
 
 		if (result.count("dequantize"))
-			dequantize = result["dequantize"].as<bool>();
+			_dequantize = result["dequantize"].as<bool>();
 
 		if (result.count("qp"))
-			qp = result["qp"].as<uint32_t>();
+			_qp = result["qp"].as<uint32_t>();
 
 		if (result.count("qt"))
-			qt = result["qt"].as<uint32_t>();
+			_qt = result["qt"].as<uint32_t>();
 
 		if (result.count("qn"))
-			qn = result["qn"].as<uint32_t>();
+			_qn = result["qn"].as<uint32_t>();
 
 		if (result.count("qc"))
-			qc = result["qc"].as<uint32_t>();
+			_qc = result["qc"].as<uint32_t>();
 
 		if (result.count("minPos")) {
-			minPosStr = result["minPos"].as<std::string>();
-			if (!parseVec3(minPosStr, minPos)) {
-				std::cout << "Error: parsing --minPos=\"" << minPosStr 
+			_minPosStr = result["minPos"].as<std::string>();
+			if (!parseVec3(_minPosStr, _minPos)) {
+				std::cout << "Error: parsing --minPos=\"" << _minPosStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("maxPos")) {
-			maxPosStr = result["maxPos"].as<std::string>();
-			if (!parseVec3(maxPosStr, maxPos)) {
-				std::cout << "Error: parsing --maxPos=\"" << maxPosStr 
+			_maxPosStr = result["maxPos"].as<std::string>();
+			if (!parseVec3(_maxPosStr, _maxPos)) {
+				std::cout << "Error: parsing --maxPos=\"" << _maxPosStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("minUv")) {
-			minUvStr = result["minUv"].as<std::string>();
-			if (!parseVec2(minUvStr, minUv)) {
-				std::cout << "Error: parsing --minUv=\"" << minUvStr 
+			_minUvStr = result["minUv"].as<std::string>();
+			if (!parseVec2(_minUvStr, _minUv)) {
+				std::cout << "Error: parsing --minUv=\"" << _minUvStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("maxUv")) {
-			maxUvStr = result["maxUv"].as<std::string>();
-			if (!parseVec2(maxUvStr, maxUv)) {
-				std::cout << "Error: parsing --maxUv=\"" << maxUvStr 
+			_maxUvStr = result["maxUv"].as<std::string>();
+			if (!parseVec2(_maxUvStr, _maxUv)) {
+				std::cout << "Error: parsing --maxUv=\"" << _maxUvStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("minNrm")) {
-			minNrmStr = result["minNrm"].as<std::string>();
-			if (!parseVec3(minNrmStr, minNrm)) {
-				std::cout << "Error: parsing --minNrm=\"" << minNrmStr 
+			_minNrmStr = result["minNrm"].as<std::string>();
+			if (!parseVec3(_minNrmStr, _minNrm)) {
+				std::cout << "Error: parsing --minNrm=\"" << _minNrmStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("maxNrm")) {
-			maxNrmStr = result["maxNrm"].as<std::string>();
-			if (!parseVec3(maxNrmStr, maxNrm)) {
-				std::cout << "Error: parsing --maxNrm=\"" << maxNrmStr 
+			_maxNrmStr = result["maxNrm"].as<std::string>();
+			if (!parseVec3(_maxNrmStr, _maxNrm)) {
+				std::cout << "Error: parsing --maxNrm=\"" << _maxNrmStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("minCol")) {
-			minColStr = result["minCol"].as<std::string>();
-			if (!parseVec3(minColStr, minCol)) {
-				std::cout << "Error: parsing --minCol=\"" << minColStr
+			_minColStr = result["minCol"].as<std::string>();
+			if (!parseVec3(_minColStr, _minCol)) {
+				std::cout << "Error: parsing --minCol=\"" << _minColStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
 		}
 		if (result.count("maxCol")) {
-			maxColStr = result["maxCol"].as<std::string>();
-			if (!parseVec3(maxColStr, maxCol)) {
-				std::cout << "Error: parsing --maxCol=\"" << maxColStr
+			_maxColStr = result["maxCol"].as<std::string>();
+			if (!parseVec3(_maxColStr, _maxCol)) {
+				std::cout << "Error: parsing --maxCol=\"" << _maxColStr
 					<< "\" expected three floats with space separator" << std::endl;
 				return false;
 			}
+		}
+		if (result.count("useFixedPoint")) {
+			_useFixedPoint = result["useFixedPoint"].as<bool>();
 		}
 	}
 	catch (const cxxopts::OptionException& e)
@@ -206,11 +211,11 @@ bool Quantize::process(uint32_t frame) {
 
 	// the input
 	Model* inputModel;
-	if ((inputModel = IO::loadModel(inputModelFilename)) == NULL) {
+	if ((inputModel = IO::loadModel(_inputModelFilename)) == NULL) {
 		return false;
 	}
 	if (inputModel->vertices.size() == 0) {
-		std::cout << "Error: invalid input model from " << inputModelFilename << std::endl;
+		std::cout << "Error: invalid input model from " << _inputModelFilename << std::endl;
 		return false;
 	}
 
@@ -221,34 +226,34 @@ bool Quantize::process(uint32_t frame) {
 	clock_t t1 = clock();
 
 	std::cout << "Quantizing" << std::endl;
-	std::cout << "  qp = " << qp << std::endl;
-	std::cout << "  qt = " << qt << std::endl;
-	std::cout << "  qn = " << qn << std::endl;
-	std::cout << "  qc = " << qn << std::endl;
-	
-	if (dequantize) {
-		Model* quantizedModel = new Model();
-		
-		quantize(*inputModel, *quantizedModel, qp, qt, qn, qc, 
-			minPos, maxPos, minUv, maxUv, minNrm, maxNrm, minCol, maxCol,
-			outputVarFilename );
+	std::cout << "  qp = " << _qp << std::endl;
+	std::cout << "  qt = " << _qt << std::endl;
+	std::cout << "  qn = " << _qn << std::endl;
+	std::cout << "  qc = " << _qn << std::endl;
 
-		Dequantize::dequantize(*quantizedModel, *outputModel, qp, qt, qn, qc, 
-			minPos, maxPos, minUv, maxUv, minNrm, maxNrm, minCol, maxCol);
-		
+	if (_dequantize) {
+		Model* quantizedModel = new Model();
+
+		quantize(*inputModel, *quantizedModel, _qp, _qt, _qn, _qc,
+			_outputVarFilename, _useFixedPoint,
+			_minPos, _maxPos, _minUv, _maxUv, _minNrm, _maxNrm, _minCol, _maxCol);
+
+		Dequantize::dequantize(*quantizedModel, *outputModel, _qp, _qt, _qn, _qc,
+			_minPos, _maxPos, _minUv, _maxUv, _minNrm, _maxNrm, _minCol, _maxCol, _useFixedPoint);
+
 		delete quantizedModel;
 	}
 	else {
-		quantize(*inputModel, *outputModel, qp, qt, qn, qc, 
-			minPos, maxPos, minUv, maxUv, minNrm, maxNrm, minCol, maxCol,
-			outputVarFilename );
+		quantize(*inputModel, *outputModel, _qp, _qt, _qn, _qc,
+			_outputVarFilename, _useFixedPoint,
+			_minPos, _maxPos, _minUv, _maxUv, _minNrm, _maxNrm, _minCol, _maxCol);
 	}
-	
+
 	clock_t t2 = clock();
 	std::cout << "Time on processing: " << ((float)(t2 - t1)) / CLOCKS_PER_SEC << " sec." << std::endl;
 
 	// save the result
-	if (IO::saveModel(outputModelFilename, outputModel))
+	if (IO::saveModel(_outputModelFilename, outputModel))
 		return true;
 	else {
 		delete outputModel;
@@ -258,23 +263,28 @@ bool Quantize::process(uint32_t frame) {
 	return true;
 }
 
-//
+// input can be mesh or point cloud
+// the min/max values are not set to const so that internally computed min/max
+// will have the expected impact onto the --dequant option of the quantize method (see. Quantize::process).
 void Quantize::quantize(
-	const Model& input, Model& output, 
+	const Model& input, Model& output,
 	const uint32_t qp, const uint32_t qt, const uint32_t qn, const uint32_t qc,
-	const glm::vec3& minPos, const glm::vec3& maxPos, const glm::vec2& minUv, const glm::vec2& maxUv,
-	const glm::vec3& minNrm, const glm::vec3& maxNrm, const glm::vec3& minCol, const glm::vec3& maxCol,
-	const std::string& outputVarFilename )
+	const std::string& outputVarFilename, const bool useFixedPoint,
+	glm::vec3& minPos, glm::vec3& maxPos, glm::vec2& minUv, glm::vec2& maxUv,
+	glm::vec3& minNrm, glm::vec3& maxNrm, glm::vec3& minCol, glm::vec3& maxCol
+)
 {
-	// copy the input	
-	output = input;
-	
-	// prepare loging and output var
+	// copy the input
+	Model tmpModel = input;
+
+	// prepare logging and output var
 	std::vector<std::ostream*> out;
 	out.push_back(&std::cout);
 	std::ofstream fout;
 	if (outputVarFilename != "") {
 		fout.open(outputVarFilename.c_str(), std::ios::out);
+		// this is mandatory to print floats with full precision
+		fout.precision(std::numeric_limits< float >::max_digits10);
 		if (fout) {
 			out.push_back(&fout);
 		}
@@ -289,25 +299,47 @@ void Quantize::quantize(
 		if (minPos == maxPos) {
 			std::cout << "Computing positions range" << std::endl;
 			computeBBox(input.vertices, minBox, maxBox);
+			minPos = minBox;
+			maxPos = maxBox;
 		}
 		else {
 			std::cout << "Using parameter positions range" << std::endl;
 			minBox = minPos; maxBox = maxPos;
 		}
+		const int32_t fixedPoint16 = (1u << 16);
+		if (useFixedPoint) {
+			// converting the values to a fixed point representation
+			// minBox(FP16) will be used in AAPS -> shift
+			for (int i = 0; i < 3; i++) {
+				if (minBox[i] > 0)
+					minBox[i] = (std::floor(minBox[i] * fixedPoint16)) / fixedPoint16;
+				else
+					minBox[i] = (-1) * (std::ceil(std::abs(minBox[i]) * fixedPoint16)) / fixedPoint16;
+				if (maxBox[i] > 0)
+					maxBox[i] = (std::ceil(maxBox[i] * fixedPoint16)) / fixedPoint16;
+				else
+					maxBox[i] = (-1) * (std::floor(std::abs(maxBox[i]) * fixedPoint16)) / fixedPoint16;
+			}
+		}
 		const glm::vec3 diag = maxBox - minBox;
 		const float range = std::max(std::max(diag.x, diag.y), diag.z);
 		const int32_t maxPositionQuantizedValue = (1u << static_cast<uint32_t>(qp)) - 1;
+		double scale = (double)range / maxPositionQuantizedValue;
+		if (useFixedPoint)
+			scale = (std::ceil(scale * fixedPoint16)) / fixedPoint16;
 
 		for (size_t i = 0; i < out.size(); ++i) {
 			*out[i] << "  minPos=\"" << minBox.x << " " << minBox.y << " " << minBox.z << "\"" << std::endl;
 			*out[i] << "  maxPos=\"" << maxBox.x << " " << maxBox.y << " " << maxBox.z << "\"" << std::endl;
 			*out[i] << "  rangePos=" << range << std::endl;
+			*out[i] << "  maxPositionQuantizedValue=" << maxPositionQuantizedValue << std::endl;
+			*out[i] << "  scale=" << scale << std::endl;
 		}
 
 		for (size_t i = 0; i < input.vertices.size() / 3; i++) {
 			for (glm::vec3::length_type c = 0; c < 3; ++c) {
-				uint32_t pos = static_cast<uint32_t> (std::floor(((input.vertices[i * 3 + c] - minBox[c]) / range) * maxPositionQuantizedValue + 0.5f));
-				output.vertices[i * 3 + c] = static_cast<float> (pos);
+				uint32_t pos = static_cast<uint32_t> (std::floor(((double(input.vertices[i * 3 + c] - minBox[c])) / scale) + 0.5f));
+				tmpModel.vertices[i * 3 + c] = static_cast<float> (pos);
 			}
 		}
 	}
@@ -318,6 +350,8 @@ void Quantize::quantize(
 		if (minUv == maxUv) {
 			std::cout << "Computing uv coordinates range" << std::endl;
 			computeBBox(input.uvcoords, minBox, maxBox);
+			minUv = minBox;
+			maxUv = maxBox;
 		}
 		else {
 			std::cout << "Using parameter uv coordinates range" << std::endl;
@@ -326,7 +360,7 @@ void Quantize::quantize(
 		const glm::vec2 diag = maxBox - minBox;
 		const float range = std::max(diag.x, diag.y);
 		const int32_t maxUVcordQuantizedValue = (1u << static_cast<uint32_t>(qt)) - 1;
-		
+
 		for (size_t i = 0; i < out.size(); ++i) {
 			*out[i] << "  minUv=\"" << minBox.x << " " << minBox.y << "\"" << std::endl;
 			*out[i] << "  maxUv=\"" << maxBox.x << " " << maxBox.y << "\"" << std::endl;
@@ -336,7 +370,7 @@ void Quantize::quantize(
 		for (size_t i = 0; i < input.uvcoords.size() / 2; i++) {
 			for (glm::vec2::length_type c = 0; c < 2; ++c) {
 				uint32_t uv = static_cast<uint32_t> (std::floor(((input.uvcoords[i * 2 + c] - minBox[c]) / range) * maxUVcordQuantizedValue + 0.5f));
-				output.uvcoords[i * 2 + c] = static_cast<float> (uv);
+				tmpModel.uvcoords[i * 2 + c] = static_cast<float> (uv);
 			}
 		}
 	}
@@ -347,6 +381,8 @@ void Quantize::quantize(
 		if (minNrm == maxNrm) {
 			std::cout << "Computing normals range" << std::endl;
 			computeBBox(input.normals, minBox, maxBox);
+			minNrm = minBox;
+			maxNrm = maxBox;
 		}
 		else {
 			std::cout << "Using parameter normals range" << std::endl;
@@ -355,7 +391,7 @@ void Quantize::quantize(
 		const glm::vec3 diag = maxBox - minBox;
 		const float range = std::max(std::max(diag.x, diag.y), diag.z);
 		const int32_t maxNormalQuantizedValue = (1u << static_cast<uint32_t>(qn)) - 1;
-		
+
 		for (size_t i = 0; i < out.size(); ++i) {
 			*out[i] << "  minNrm=\"" << minBox.x << " " << minBox.y << " " << minBox.z << "\"" << std::endl;
 			*out[i] << "  maxNrm=\"" << maxBox.x << " " << maxBox.y << " " << maxBox.z << "\"" << std::endl;
@@ -365,7 +401,7 @@ void Quantize::quantize(
 		for (size_t i = 0; i < input.normals.size() / 3; i++) {
 			for (glm::vec3::length_type c = 0; c < 3; ++c) {
 				uint32_t nrm = static_cast<uint32_t> (std::floor(((input.normals[i * 3 + c] - minBox[c]) / range) * maxNormalQuantizedValue + 0.5f));
-				output.normals[i * 3 + c] = static_cast<float> (nrm);
+				tmpModel.normals[i * 3 + c] = static_cast<float> (nrm);
 			}
 		}
 	}
@@ -376,6 +412,8 @@ void Quantize::quantize(
 		if (minCol == maxCol) {
 			std::cout << "Computing colors range" << std::endl;
 			computeBBox(input.normals, minBox, maxBox);
+			minCol = minBox;
+			maxCol = maxBox;
 		}
 		else {
 			std::cout << "Using parameter colors range" << std::endl;
@@ -384,7 +422,7 @@ void Quantize::quantize(
 		const glm::vec3 diag = maxBox - minBox;
 		const float range = std::max(std::max(diag.x, diag.y), diag.z);
 		const int32_t maxColorQuantizedValue = (1u << static_cast<uint32_t>(qn)) - 1;
-		
+
 		for (size_t i = 0; i < out.size(); ++i) {
 			*out[i] << "  minCol=\"" << minBox.x << " " << minBox.y << " " << minBox.z << "\"" << std::endl;
 			*out[i] << "  maxCol=\"" << maxBox.x << " " << maxBox.y << " " << maxBox.z << "\"" << std::endl;
@@ -394,8 +432,55 @@ void Quantize::quantize(
 		for (size_t i = 0; i < input.colors.size() / 3; i++) {
 			for (glm::vec3::length_type c = 0; c < 3; ++c) {
 				uint32_t col = static_cast<uint32_t> (std::floor(((input.colors[i * 3 + c] - minBox[c]) / range) * maxColorQuantizedValue + 0.5f));
-				output.colors[i * 3 + c] = static_cast<float> (col);
+				tmpModel.colors[i * 3 + c] = static_cast<float> (col);
 			}
 		}
 	}
+
+	// not a mesh we do not need to cleanup
+	// copy is not optimal, shall better use pointers at the begining of the func
+	if (tmpModel.triangles.size() == 0) {
+		output = tmpModel;
+		return;
+	}
+
+	// remove degenerate triangles, and use common index for all attributes
+	std::cout << "Cleanup quantized mesh " << std::endl;
+
+	// preserve materials or other infos
+	output.comments = tmpModel.comments;
+	output.header = tmpModel.header;
+
+	// to prevent storing duplicate points, we use a ModelBuilder
+	ModelBuilder builder(output);
+
+	size_t skipped = 0; // number of degenerate triangles
+
+	// For each triangle
+	for (size_t triIdx = 0; triIdx < tmpModel.triangles.size() / 3; ++triIdx) {
+
+		Vertex v1, v2, v3;
+
+		fetchTriangle(tmpModel, triIdx,
+			input.uvcoords.size() != 0,
+			input.colors.size() != 0,
+			input.normals.size() != 0,
+			v1, v2, v3);
+
+		// check if triangle is not degenerate
+		if (triangleArea(v1.pos, v2.pos, v3.pos) < DBL_EPSILON) {
+			++skipped;
+			continue;
+		}
+
+		// add triangle to output and reindex
+		builder.pushTriangle(v1, v2, v3);
+
+	}
+
+	if (skipped != 0)
+		std::cout << "Skipped " << skipped << " degenerate triangles" << std::endl;
+	if (builder.foundCount != 0)
+		std::cout << "Handled " << builder.foundCount << " duplicate vertices" << std::endl;
+	std::cout << "Generated " << output.vertices.size() / 3 << " points" << std::endl;
 }