+++ source/mmQuantize.h
@@ -44,46 +44,47 @@ public:
 private:
 
 	// Command parameters
-	std::string inputModelFilename;
-	std::string outputModelFilename;
-	std::string outputVarFilename;
+	std::string _inputModelFilename;
+	std::string _outputModelFilename;
+	std::string _outputVarFilename;
 	// Quantization parameters
-	uint32_t qp = 12;	// geometry
-	uint32_t qt = 12;	// UV coordinates
-	uint32_t qn = 12;	// normals
-	uint32_t qc = 8;	// colors
-	bool dequantize = false;
+	uint32_t _qp = 12;	// geometry
+	uint32_t _qt = 12;	// UV coordinates
+	uint32_t _qn = 12;	// normals
+	uint32_t _qc = 8;	// colors
+	bool _dequantize = false;
 	//
-	bool computeBboxPos = false;
-	bool computeBboxUv = false;
-	bool computeBboxNrm = false;
-	bool computeBboxCol = false;
+	bool _computeBboxPos = false;
+	bool _computeBboxUv = false;
+	bool _computeBboxNrm = false;
+	bool _computeBboxCol = false;
+	bool _useFixedPoint = false;
 	// min max vectors
-	std::string minPosStr;
-	std::string maxPosStr;
-	glm::vec3 minPos = { 0.0F,0.0F,0.0F };
-	glm::vec3 maxPos = { 0.0F,0.0F,0.0F };
-	std::string minUvStr;
-	std::string maxUvStr;
-	glm::vec2 minUv = { 0.0F,0.0F };
-	glm::vec2 maxUv = { 0.0F,0.0F };
-	std::string minNrmStr;
-	std::string maxNrmStr;
-	glm::vec3 minNrm = { 0.0F,0.0F,0.0F };
-	glm::vec3 maxNrm = { 0.0F,0.0F,0.0F };
-	std::string minColStr;
-	std::string maxColStr;
-	glm::vec3 minCol = { 0.0F,0.0F,0.0F };
-	glm::vec3 maxCol = { 0.0F,0.0F,0.0F };
-	
+	std::string _minPosStr;
+	std::string _maxPosStr;
+	glm::vec3 _minPos = { 0.0F,0.0F,0.0F };
+	glm::vec3 _maxPos = { 0.0F,0.0F,0.0F };
+	std::string _minUvStr;
+	std::string _maxUvStr;
+	glm::vec2 _minUv = { 0.0F,0.0F };
+	glm::vec2 _maxUv = { 0.0F,0.0F };
+	std::string _minNrmStr;
+	std::string _maxNrmStr;
+	glm::vec3 _minNrm = { 0.0F,0.0F,0.0F };
+	glm::vec3 _maxNrm = { 0.0F,0.0F,0.0F };
+	std::string _minColStr;
+	std::string _maxColStr;
+	glm::vec3 _minCol = { 0.0F,0.0F,0.0F };
+	glm::vec3 _maxCol = { 0.0F,0.0F,0.0F };
+
 	// quantizes the model using parameters and range. If a range is not set (i.e. min == max) then a range is 
 	// computed internally. If quantization parameter is < 7, the related attribute is not quantized.
 	static void quantize(
 		const Model& input, Model& output,
 		const uint32_t qp, const uint32_t qt, const uint32_t qn, const uint32_t qc,
-		const glm::vec3& minPos, const glm::vec3& maxPos, const glm::vec2& minUv, const glm::vec2& maxUv,
-		const glm::vec3& minNrm, const glm::vec3& maxNrm, const glm::vec3& minCol, const glm::vec3& maxCol,
-		const std::string& outputVarFilename);
+		const std::string& outputVarFilename, bool useFixedPoint,
+		glm::vec3& minPos, glm::vec3& maxPos, glm::vec2& minUv, glm::vec2& maxUv,
+		glm::vec3& minNrm, glm::vec3& maxNrm, glm::vec3& minCol, glm::vec3& maxCol);
 
 };